Convert Videoheader component to TypeScript
